% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_crown.R
\name{classify_crown}
\alias{classify_crown}
\title{Classify tree crown and trunk points in a point cloud}
\usage{
classify_crown(file, res = 0.1, denoise = TRUE, max_crown_height = 0.4)
}
\arguments{
\item{file}{las/laz file or lidR object. A point cloud file of an individual tree
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.}

\item{res}{Numeric. Resolution/intervals (i.e. height in meters) at which to slice the point cloud.}

\item{denoise}{Logical. Determines whether point cloud noise should be removed (determined by lidR::classify_noise).}

\item{max_crown_height}{Numeric. The maximum height at which the crown can start (presented as proportional height).
Default is 0.4 (i.e. the crown must start =< 40\% up the tree).}
}
\value{
list of two point clouds, the first being the tree crown and the second being the trunk
}
\description{
Separates a tree point cloud into the crown and trunk.
}
