% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_complexity.R
\name{compute_curvature}
\alias{compute_curvature}
\title{Per-point curvatures}
\usage{
compute_curvature(points, k = 20)
}
\arguments{
\item{points}{lidR object. A point cloud file of an object.
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.}

\item{k}{Numeric. Number of nearest neighbors to be considered when computing curvatures.
Default is 20.}
}
\value{
Numeric. A vector of numbers.
}
\description{
The coherence of each point as considered by its 20 nearest neighbors.
Based on the eigenstructure method of computing 3D structural coherence. see Gersztenkorn and Marfurt 1999.
This function is meant to be used in the crown_curvature function.
}
