% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_complexity.R
\name{crown_curvature}
\alias{crown_curvature}
\title{Crown curvature}
\usage{
crown_curvature(crown_pc, k = 20)
}
\arguments{
\item{crown_pc}{lidR object. A point cloud file of an individual tree crown.
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.
Tree crown point cloud can be obtained from the classify_crown function.}

\item{k}{Numeric. Number of nearest neighbors to be considered when computing curvatures.
Default is 20.}
}
\value{
List of 3 numeric values corresponding to the mean, standard deviation,
and 90th percentile curvature values, respectively.
}
\description{
The mean, standard deviation, and 90th percentile crown curvature
as computed by compute_curvature when applied to crown_pc.
}
