% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_complexity.R
\name{crown_fullness}
\alias{crown_fullness}
\title{Crown fullness}
\usage{
crown_fullness(crown_pc, tree_pc)
}
\arguments{
\item{crown_pc}{lidR object. A point cloud file of an individual tree crown.
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.
Tree crown point cloud can be obtained from the classify_crown function.}

\item{tree_pc}{lidR object. A point cloud file of an individual tree
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.}
}
\value{
Numeric. A proportion (values range from 0 to 1)
}
\description{
Proportion of points located in the crown compared to the entire tree,
scaled by the height of the crown. ALternatively, a measure of crown density.
}
