% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_crown_dimensions.R
\name{crown_height}
\alias{crown_height}
\title{Crown height}
\usage{
crown_height(crown_pc)
}
\arguments{
\item{crown_pc}{lidR object. A point cloud file of an individual tree crown.
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.
Tree crown point cloud can be obtained from the classify_crown function.}
}
\value{
Numeric. Crown height (units match units of the point cloud X, Y, Z coordinates)
}
\description{
Calculates crown height as the difference between the minimum and maximum Z values of the
tree crown's point cloud.
}
