% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_crown_dimensions.R
\name{crown_projected_area}
\alias{crown_projected_area}
\title{Crown projected areaList containing 3 numeric values corresponding to the crown width at the top, middle, and bottom of the crown, respectively.}
\usage{
crown_projected_area(crown_pc, res = 0.1, npts_empty = 20)
}
\arguments{
\item{crown_pc}{lidR object. A point cloud file of an individual tree crown.
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.
Tree crown point cloud can be obtained from the classify_crown function.}

\item{res}{Numeric. Resolution of the voxels used to calculate crown projected area and gap fraction.
Default is 0.1 meters.}

\item{npts_empty}{Numeric. Minimum number of points required for a voxel to be considered not empty (i.e. containing vegetation).
Default is 20 points. Value used in crown projected area and gap fraction calculations.}
}
\value{
Numeric. A proportion (values range from 0 to 1).
}
\description{
Calculates crown projected area as the area the crown would occupy if lain flat on the ground.
In other words, the 2D area of the crown as seen from above. Calculated by voxelating the crown,
projecting it into 2D space, and multiplying the number of pixels by the resolution of the voxels.
}
