% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_crown_dimensions.R
\name{crown_shape}
\alias{crown_shape}
\title{Crown shape}
\usage{
crown_shape(
  crown_pc,
  buffer = 0.1,
  slice_radius = 0.1,
  top_height = 0.9,
  middle_height = 0.5,
  bottom_height = 0.1
)
}
\arguments{
\item{crown_pc}{lidR object. A point cloud file of an individual tree crown.
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.
Tree crown point cloud can be obtained from the classify_crown function.}

\item{buffer}{Numeric. Proportion of the crown widths (value ranges from 0 to 1). The permissable buffer zone before
the function determines whether one width is smaller of larger than another.}

\item{slice_radius}{Numeric. Half the height in meters of the slices used to calculate crown widths.
Default is 0.1 meters.}

\item{top_height}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the top of the crown
will be calculated. Default is 0.9.}

\item{middle_height}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the middle of the crown
will be calculated. Default is 0.5.}

\item{bottom_height}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the bottom of the crown
will be calculated. Default is 0.1.}
}
\value{
Character.
}
\description{
Calculates crown shape based on the crown widths at the top, middle, and bottom of the crown
(as computed by the function crown_width). Returns crown shape as one of four options: conal (where bottom
crown width is the smallest, followed by the middle and the top), vase (where bottom crown width is the
largest, followed by the middle and the top), column (where all widths are of the same length, allowing
for a buffer zone), and oval (where the middle width is the largest).
Since there are only four shapes, this is a rough estimate (i.e. oval and round shapes are essentially
the same here).
The function therefore first computes crown_widths then determines crown shape from these values.
}
