% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_crown_dimensions.R
\name{crown_width}
\alias{crown_width}
\title{Crown width/diameter}
\usage{
crown_width(
  crown_pc,
  slice_radius = 0.1,
  top_height = 0.9,
  middle_height = 0.5,
  bottom_height = 0.1
)
}
\arguments{
\item{crown_pc}{lidR object. A point cloud file of an individual tree crown.
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.
Tree crown point cloud can be obtained from the classify_crown function.}

\item{slice_radius}{Numeric. Half the height in meters of the slices used to calculate crown widths.
Default is 0.1 meters.}

\item{top_height}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the top of the crown
will be calculated. Default is 0.9.}

\item{middle_height}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the middle of the crown
will be calculated. Default is 0.5.}

\item{bottom_height}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the bottom of the crown
will be calculated. Default is 0.1.}
}
\value{
List containing 3 numeric values corresponding to the crown width at the top, middle, and bottom of the crown, respectively.
}
\description{
Crown width is calculated by calculating crown radius at 5° intervals,
computing the mean, and multiplying this mean by 2.
}
\details{
Details of the calculation are as follows:
(1) Isolated the middle of the crown by slicing the crown at its midpoint (at 50\% of its maximum height),
keeping 10 cm above and below for a total slice thickness of 20 cm. (2) Measured the radial distance
from the tree’s central axis to the furthest point in the sliced crown. (3) Repeated this measure in
5° intervals along the center axis and calculated the average. (4) Multiplied this average by two to
estimate crown diameter.
}
