% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{extract_features}
\alias{extract_features}
\title{Extraction of tree architectural features}
\usage{
extract_features(
  file,
  res = 0.1,
  denoise = TRUE,
  max_crown_height = 0.4,
  res_vox = 0.1,
  npts_empty = 20,
  slice_radius = 0.1,
  top_crown = 0.9,
  middle_crown = 0.5,
  bottom_crown = 0.1,
  top_trunk = 0.6,
  middle_trunk = 0.5,
  bottom_trunk = 0.4,
  k = 20
)
}
\arguments{
\item{file}{las/laz file or lidR object. A point cloud file of an individual tree
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.}

\item{res}{Numeric. Resolution/intervals (i.e. height in meters) at which to slice the point cloud (for classifying crown points).}

\item{denoise}{Logical. Determines whether point cloud noise should be removed (determined by lidR::classify_noise).}

\item{max_crown_height}{Numeric. The maximum height at which the crown can start (presented as proportional height).
Default is 0.4 (i.e. the crown must start =< 40\% up the tree).}

\item{res_vox}{Numeric. Resolution of the voxels used to calculate crown projected area and gap fraction.
Default is 0.1 meters.}

\item{npts_empty}{Numeric. Minimum number of points required for a voxel to be considered not empty (i.e. containing vegetation).
Default is 20 points. Value used in crown projected area and gap fraction calculations.}

\item{slice_radius}{Numeric. Half the height in meters of the slices used to calculate crown width and trunk widths.
Default is 0.1 meters.}

\item{top_crown}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the top of the crown
will be calculated. Default is 0.9.}

\item{middle_crown}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the middle of the crown
will be calculated. Default is 0.5.}

\item{bottom_crown}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the bottom of the crown
will be calculated. Default is 0.1.}

\item{top_trunk}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the top of the trunk
will be calculated. Default is 0.6.}

\item{middle_trunk}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the middle of the trunk
will be calculated. Default is 0.5.}

\item{bottom_trunk}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the bottom of the trunk
will be calculated. Default is 0.4.}

\item{k}{Numeric. Number of nearest neighbors to be considered when computing curvatures. Default is 20}
}
\value{
A dataframe containing all tree architecural features.
}
\description{
This function calculates tree archiectural features pertaining to  . In effect, it combines
all functions available in the entireity of the package and adds a few. Particularly, some
ratios of width and height of different parts of the tree (e.g. ratio of crown height to
tree height).
}
\details{
The entire list of features rendered (totaling 23) is as follows:
tree height, crown height, crown width (at the upper limit, lower limit, and middle of the
crown), trunk width, crown projected area, ratio of crown projected area to tree height,
ratio of crown height to tree height, ratio of crown width to tree height,
ratio of crown width to trunk width, ratio of upper crown width to lower crown width,
ratio of lower crown width to middle crown width, ratio of upper crown width to middle
crown width, crown fullness, gap fraction, box dimension, curvature mean, curvature standard
deviation, curvature p90, rumple index.
}
\examples{
tree <- lidR::readTLSLAS(file)
features <- extract_features(tree)
}
