% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_complexity.R
\name{gap_fraction}
\alias{gap_fraction}
\title{Gap fraction}
\usage{
gap_fraction(crown_pc, res = 0.1, npts_empty = 20)
}
\arguments{
\item{crown_pc}{lidR object. A point cloud file of an individual tree crown.
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.
Tree crown point cloud can be obtained from the classify_crown function.}

\item{res}{Numeric. Resolution of the voxels used to calculate crown projected area and gap fraction.
Default is 0.1 meters.}

\item{npts_empty}{Numeric. Minimum number of points required for a voxel to be considered not empty (i.e. containing vegetation).
Default is 20 points. Value used in crown projected area and gap fraction calculations.}
}
\value{
Numeric. A proportion (values range from 0 to 1)
}
\description{
The proportion of the 2D projected crown voxels that contain less than 20 points.
Calculated by voxelating the crown (with a default resolution of 10 cm), projecting it into 2D space,
and calculating the proportion of ‘gaps’, whereby gaps are characterized as all voxels
containing less than 20 points.
}
