% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_crown_dimensions.R
\name{tree_height}
\alias{tree_height}
\title{Tree height}
\usage{
tree_height(tree_pc)
}
\arguments{
\item{tree_pc}{lidR object. A point cloud file of an individual tree
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.}
}
\value{
Numeric. Tree height (units match units of the point cloud X, Y, Z coordinates)
}
\description{
Calculates tree height as the difference between the minimum and maximum Z values of the point cloud.
}
