% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_crown_dimensions.R
\name{trunk_width}
\alias{trunk_width}
\title{Trunk width/diameter}
\usage{
trunk_width(
  trunk_pc,
  slice_radius = 0.1,
  top_height = 0.6,
  middle_height = 0.5,
  bottom_height = 0.4
)
}
\arguments{
\item{trunk_pc}{lidR object. A point cloud file of an individual tree trunk.
Designed for single-spectral TLS (terrestrial laser scanning) or
high-resolution MLS (mobile laser scanning) data.
Tree trunk point cloud can be obtained from the classify_crown function.}

\item{slice_radius}{Numeric. Half the height in meters of the slices used to calculate the trunk width.
Default is 0.1 meters.}

\item{top_height}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the top of the trunk
will be calculated. Default is 0.6.}

\item{middle_height}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the middle of the trunk
will be calculated. Default is 0.5.}

\item{bottom_height}{Numeric. Proportional crown height (values range from 0 to 1) at which the width of the bottom of the trunk
will be calculated. Default is 0.4.}
}
\value{
Numeric. Trunk width (units match units of the point cloud X, Y, Z coordinates)
}
\description{
Trunk width is calculated by calculating crown radius at 5° intervals at different heights
along the trunk, computing the mean, and multiplying this mean by 2.
}
\details{
Equivalent to the crown width calculation but measured 10 times per trunk at 10\% height intervals.
The mean trunk width across all height intervals is taken.
}
